requirements.txt
email-validator==2.1.1
faust-streaming==0.11.0
faker<24.1.0,>=17.0.0
fastapi==0.110.0
googlemaps==4.10.0
#graphdatascience==1.9
joblib==1.3.2
loguru==0.7.2
lz4==4.3.3
neo4j==5.18.0
phonenumbers==8.13.33
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pymongo==4.6.3
pytest==8.1.1
pytest-asyncio==0.23.5.post1
python-multipart==0.0.9
python-schema-registry-client==2.5.6
redis==5.0.3
sdv==1.11.0
tomlkit==0.12.4
uvicorn==0.13.4